requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
blis==0.7.4
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
corenlp-protobuf==3.8.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.0.0/en_core_web_trf-3.0.0-py3-none-any.whl
entrypoints==0.3
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
importlib-metadata==3.7.0
inflect==5.3.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lxml==4.6.2
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.5
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pathtools==0.1.2
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
preshed==3.0.5
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.15.5
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pydantic==1.7.3
pydeck==0.6.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
sentencepiece==0.1.95
sentry-sdk==1.0.0
seqeval==1.2.2
shortuuid==1.0.1
simpletransformers==0.60.9
six==1.15.0
smart-open==3.0.0
smmap==3.0.5
spacy==3.0.5
spacy-alignments==0.8.3
spacy-legacy==3.0.1
spacy-transformers==1.0.2
srsly==2.4.0
stanford-corenlp==3.9.2
stanza==1.2
streamlit==0.78.0
subprocess32==3.5.4
tensorboardX==2.1
terminado==0.9.2
testpath==0.4.4
thinc==8.0.2
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
toolz==0.11.1
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
transformers==4.3.3
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
wandb==0.10.21
wasabi==0.8.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
word2number==1.1
zipp==3.4.1
